# docs/api/openapi_plp_v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SHIOL+ Powerball Prediction API",
    "description": "Provides ML-based Powerball number predictions.",
    "version": "6.0.0"
  },
  "paths": {
...
    "/api/v1/predictions/public/by-draw/{draw_date}": {
      "get": {
        "summary": "Get Public Predictions By Draw Date",
        "description": "Public endpoint for getting predictions by draw date (frontend access)",
        "operationId": "get_public_predictions_by_draw_date_api_v1_predictions_public_by_draw__draw_date__get",
        "parameters": [
          {
            "name": "draw_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draw Date"
            }
          },
          {
            "name": "min_matches",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Minimum number of matches required",
              "default": 0,
              "title": "Min Matches"
            },
            "description": "Minimum number of matches required"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Maximum number of predictions to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of predictions to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predictions/public/predictions-only/{draw_date}": {
      "get": {
        "summary": "Get Predictions Only By Date",
        "description": "Get predictions for a specific date without requiring official draw results",
        "operationId": "get_predictions_only_by_date_api_v1_predictions_public_predictions_only__draw_date__get",
        "parameters": [
          {
            "name": "draw_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draw Date"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Maximum number of predictions to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of predictions to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predictions/generate-multi-strategy": {
      "post": {
        "summary": "Generate Tickets Multi Strategy",
        "description": "Generate tickets using the multi-strategy system with adaptive weights.\n\nArgs:\n    count: Number of tickets to generate (1-100, default 5)\n\nReturns:\n    Dict with tickets, metadata, and strategy distribution",
        "operationId": "generate_tickets_multi_strategy_api_v1_predictions_generate_multi_strategy_post",
        "parameters": [
          {
            "name": "count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 5,
              "title": "Count"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Generate Tickets Multi Strategy Api V1 Predictions Generate Multi Strategy Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ticket/verify": {
      "post": {
        "tags": [
          "ticket"
        ],
        "summary": "Verify Ticket Image",
        "description": "Verify a Powerball ticket by processing an uploaded image.\n\nArgs:\n    request: FastAPI request object\n    file: Uploaded image file (JPG, PNG, etc.)\n    manual_date: Optional manual date override\n    \nReturns:\n    JSON response with verification results",
        "operationId": "verify_ticket_image_api_v1_ticket_verify_post",
        "parameters": [
          {
            "name": "manual_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Manual Date"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_verify_ticket_image_api_v1_ticket_verify_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ticket/preview": {
      "post": {
        "tags": [
          "ticket"
        ],
        "summary": "Preview Ticket Numbers",
        "description": "Preview numbers detected from ticket image (no limits applied).\nShows what OCR detected without consuming verification quotas.\n\nArgs:\n    request: FastAPI request object\n    file: Uploaded ticket image file\n    \nReturns:\n    Detected numbers and metadata",
        "operationId": "preview_ticket_numbers_api_v1_ticket_preview_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_preview_ticket_numbers_api_v1_ticket_preview_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ticket/verify-manual": {
      "post": {
        "tags": [
          "ticket"
        ],
        "summary": "Verify Manual Plays",
        "description": "Verify manually entered lottery numbers against official results.\n\nArgs:\n    fastapi_request: FastAPI request object\n    request: Manual verification request with plays and draw date\n    \nReturns:\n    JSON response with verification results",
        "operationId": "verify_manual_plays_api_v1_ticket_verify_manual_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/public/predictions-only/{draw_date}": {
      "get": {
        "tags": [
          "plp_v2"
        ],
        "summary": "Plp Predictions Only",
        "description": "Return predictions for a date without requiring official results (PLP v2 shape).",
        "operationId": "plp_predictions_only_api_v2_public_predictions_only__draw_date__get",
        "parameters": [
          {
            "name": "draw_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target draw date in YYYY-MM-DD",
              "title": "Draw Date"
            },
            "description": "Target draw date in YYYY-MM-DD"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Plp Predictions Only Api V2 Public Predictions Only  Draw Date  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_preview_ticket_numbers_api_v1_ticket_preview_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_preview_ticket_numbers_api_v1_ticket_preview_post"
      },
      "Body_verify_ticket_image_api_v1_ticket_verify_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_verify_ticket_image_api_v1_ticket_verify_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManualPlay": {
        "properties": {
          "line": {
            "type": "integer",
            "title": "Line"
          },
          "main_numbers": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Main Numbers"
          },
          "powerball": {
            "type": "integer",
            "title": "Powerball"
          }
        },
        "type": "object",
        "required": [
          "line",
          "main_numbers",
          "powerball"
        ],
        "title": "ManualPlay"
      },
      "ManualVerificationRequest": {
        "properties": {
          "plays": {
            "items": {
              "$ref": "#/components/schemas/ManualPlay"
            },
            "type": "array",
            "title": "Plays"
          },
          "draw_date": {
            "type": "string",
            "title": "Draw Date"
          }
        },
        "type": "object",
        "required": [
          "plays",
          "draw_date"
        ],
        "title": "ManualVerificationRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}